feat(utils): port tabListener to typescript
affects: @fremtind/jkl-utils
