+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 import { StoryTemplate } from "./internal/StoryTemplate";
+import { initTabListener } from "./tabListener";
 
-export { StoryTemplate };
+export { StoryTemplate, initTabListener };
